requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
django-debug-toolbar==4.2.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
Faker==19.3.1
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
